Adds basic unit features
- Introduces the class `Quantity` which is a composite of a float and a unit.
- Units must implement the `UnitInterface` and inherit from `UnitBase` which does some magic, making the specific unit classes more a configuration.
- Introduces a class `Calculator` which must be used to calculate with quantities. Does only support adding as of now.

All quantities store the value in the base unit of the corresponding unit class. So far, there are implementations for amount (1, but also %, ppm),
molar amount (1 mol) and mass (1 g). molar amount supports interconversion from and to amount using avogadros number.

Conversion in this subpackage means to convert within the unit type, interconversion means to convert from one to the other unit type.