Small unit rebasing and recipe API with adjustable volume
The recipe extension (called 'pole', a combination of 'php' and 'mole' ...) was reorganized in subdirectories.
Implementation of Units are now in App/Pole/Unit, the exceptions in App/Pole/Exception.
A brand new abstract class `CompositeUnit` was created to make support of composite units more flexible. The system is not perfect and needs a bunch of additional tests to force good behaviour, but works for now.
Based on this, a new `MolarConcentration` unit was created and which is based on CompositeUnit.

The recipe API was created (api/public/recipe/{recipeId}) which returns the desired recipe in JSON. Using the query string, adjusting the desired volume and the concentration factor is possible.