Cells are now grouped under CellGroups
A major update on the structure of cells to prepare them for group management introduces cell groups. Cell groups are meant to organise cells on a hierarchical level. For example, a organisatorial tree with Mammalian cells > Human cells > HEK293 > HEK293T > Flp-In-293 would be possible.
Cell groups can currently only be made by admins, but it is meant for the future to have a designated "cell librarian" role to manage this and prevent a wild west. In contrast, cells should only have a child/parent relationship if the specific cell lines have been physically derived from
each other. If you bought HEK293T and Flp-In-293 cells, they should be connected on their group level, but not on their cell line level. However, if you would make a MGMT overexpression cell line from a bought HEK293T, then that is fitting for child/parent.

Cells (but not cell groups) also have now an owner, a (owner) group and a privacy level. The privacy level determines the visibility: Public means 'everyone can see it', Group means 'everyone from the group can see it', and Private means 'only the owner can see it'.
Changing the cell's information is currently always possible if you are a member of the owning group (or an admin ...) - technically even the cell is set to private. But as you cannot see it, you cannot change it anyway.

The database migration should be as painless as possible. Manual sorting of the cell groups is necessary, though.