Adds the possibility to automatically reference conditions.
Experimental design fields can now be marked as "referenced". "Referenced" fields are used to retrieve conditions with matching fields to be used as a reference. Additionally, a "referenceValue" can be used to search for specific values inside the reference - such as a concentration of 0.

The reference can be referenced in model parameters (using fit), accessible are all fields that are accessible (floats and integers) and additionally model-derived fields.

Models can also set a reference model. The references reference model is displayed together with the data.